Profile-Management/ProfileManagementApp/src/main/java/edu/fiu/wfan/repository/UserRepository.java
package edu.fiu.wfan.repository;

import java.util.List;
import java.util.Optional;

import org.springframework.boot.autoconfigure.data.web.SpringDataWebProperties.Pageable;
import org.springframework.data.domain.Page;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

import edu.fiu.wfan.entity.User;

public interface UserRepository extends JpaRepository<User, Long> {
	
	@Query("SELECT u FROM User u WHERE u.email = ?1")
	public User findByEmail(@Param("email")String email);  //match login info with db
		
	@Query("SELECT u FROM User u WHERE u.email LIKE %?1%")
	public List<User> search(String keyword); //search user by username(email)
	
}
